+++ docs/app/documentation/component-docs/calendar/calendar-docs.component.ts
@@ -7,6 +7,8 @@ import * as calendarIntlSrc from '!raw-loader!./examples/calendar-i18n-example.c
 import * as calendarIntlMomentSrc from '!raw-loader!./examples/calendar--i18n-moment-example.component.ts';
 import * as calendarSingleFormSource from '!raw-loader!./examples/calendar-form-single-example.component.ts';
 import * as calendarRangeFormSource from '!raw-loader!./examples/calendar-form-range-example.component.ts';
+import * as calendarFormSourceT from '!raw-loader!./examples/calendar-form-examples.component.ts';
+import * as calendarFormSourceH from '!raw-loader!./examples/calendar-form-examples.component.html';
 import * as calendarProgrammaticallySource from '!raw-loader!./examples/calendar-programmatically-change-example.component.ts';
 import { ExampleFile } from '../../core-helpers/code-example/example-file';
 
@@ -104,6 +106,15 @@ myDisableFunction = function(d: FdDate): boolean {
         language: 'typescript',
         code: calendarRangeFormSource
     }];
+    calendarFormSource: ExampleFile[] = [
+        {
+            language: 'typescript',
+            code: calendarFormSourceT
+        }, {
+            language: 'html',
+            code: calendarFormSourceH
+        }
+    ];
 
     calendarProgrammaticallySource: ExampleFile[] = [{
         language: 'typescript',