+++ docs/app/documentation/component-docs/input-group/input-group-docs.component.ts
@@ -10,8 +10,6 @@ import * as inputGroupTextSrc from '!raw-loader!./examples/input-group-text-exam
 import * as inputGroupTextCompactSrc from '!raw-loader!./examples/input-group-text-compact-example.component.html';
 import * as formInputTsSrc from '!raw-loader!./examples/input-group-form-example.component.ts';
 import * as formInputHtmlSrc from '!raw-loader!./examples/input-group-form-example.component.html';
-import * as formInputNumberTsSrc from '!raw-loader!./examples/input-group-number-form-example.component.ts'
-import * as formInputNumberHtmlSrc from '!raw-loader!./examples/input-group-number-form-example.component.html'
 import { ExampleFile } from '../../core-helpers/code-example/example-file';
 
 @Component({
@@ -739,17 +737,6 @@ export class InputGroupDocsComponent {
         {
             language: 'typescript',
             code: formInputTsSrc
-        }
-    ];
-
-    formNumberInput: ExampleFile[] = [
-        {
-            language: 'html',
-            code: formInputNumberHtmlSrc
-        },
-        {
-            language: 'typescript',
-            code: formInputNumberTsSrc
         }
     ];