+++ library/src/lib/combobox/combobox.component.ts
@@ -20,6 +20,7 @@ import { ComboboxItem } from './combobox-item';
 import { MenuKeyboardService } from '../menu/menu-keyboard.service';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
+import focusTrap, { FocusTrap } from 'focus-trap';
 
 /**
  * Allows users to filter through results and select a value.
@@ -66,6 +67,14 @@ export class ComboboxComponent implements ControlValueAccessor, OnInit, OnChange
     /** Icon to display in the right-side button. */
     @Input()
     glyph: string = 'navigation-down-arrow';
+
+    /**
+     *  The trigger events that will open/close the options popover, by default it is click, so if user click on
+     *  input field, the popover with options will open or close
+     *  Accepts any [HTML DOM Events](https://www.w3schools.com/jsref/dom_obj_event.asp).
+     */
+    @Input()
+    triggers: string[] = ['click'];
 
     /**
      * The template with which to display the individual listed items.
@@ -127,21 +136,28 @@ export class ComboboxComponent implements ControlValueAccessor, OnInit, OnChange
     inputTextValue: string;
 
     /** @hidden */
+    public focusTrap: FocusTrap;
+
+    /** @hidden */
     private readonly onDestroy$: Subject<void> = new Subject<void>();
 
     /** @hidden */
-    onChange: any = () => {};
+    onChange: any = () => { };
 
     /** @hidden */
-    onTouched: any = () => {};
+    onTouched: any = () => { };
 
-    constructor(private menuKeyboardService: MenuKeyboardService) {}
+    constructor(
+        private elRef: ElementRef,
+        private menuKeyboardService: MenuKeyboardService
+    ) { }
 
     /** @hidden */
     ngOnInit() {
         if (this.dropdownValues) {
             this.displayedValues = this.dropdownValues;
         }
+        this.setupFocusTrap();
     }
 
     /** @hidden */
@@ -166,7 +182,7 @@ export class ComboboxComponent implements ControlValueAccessor, OnInit, OnChange
             .pipe(takeUntil(this.onDestroy$))
             .subscribe(index => this.onMenuClickHandler(index));
         this.menuKeyboardService.focusEscapeBeforeList = () => this.searchInputElement.nativeElement.focus();
-        this.menuKeyboardService.focusEscapeAfterList = () => {};
+        this.menuKeyboardService.focusEscapeAfterList = () => { };
     }
 
     /** @hidden */
@@ -182,9 +198,14 @@ export class ComboboxComponent implements ControlValueAccessor, OnInit, OnChange
     }
 
     /** @hidden */
-    onInputKeyupHandler() {
-        if (this.inputText && this.inputText.length) {
+    onInputKeyupHandler(event: KeyboardEvent) {
+        if (this.inputText &&
+            this.inputText.length &&
+            event.code !== 'Escape' &&
+            event.code !== 'Space' &&
+            event.code !== 'Enter') {
             this.isOpen = true;
+            this.isOpenChangeHandle(this.isOpen);
         }
     }
 
@@ -241,6 +262,21 @@ export class ComboboxComponent implements ControlValueAccessor, OnInit, OnChange
         }
     }
 
+    /** @hidden */
+    isOpenChangeHandle(isOpen: boolean): void {
+        this.isOpen = isOpen;
+        if (isOpen) {
+            this.focusTrap.activate();
+        } else {
+            this.focusTrap.deactivate();
+        }
+    }
+
+    /** @hidden */
+    setDisabledState(isDisabled: boolean): void {
+        this.disabled = isDisabled;
+    }
+
     private defaultDisplay(str: any): string {
         return str;
     }
@@ -257,10 +293,25 @@ export class ComboboxComponent implements ControlValueAccessor, OnInit, OnChange
     private handleClickActions(term): void {
         if (this.closeOnSelect) {
             this.isOpen = false;
+            this.isOpenChangeHandle(this.isOpen);
         }
         if (this.fillOnSelect) {
             this.inputText = this.displayFn(term);
             this.handleSearchTermChange();
+        }
+    }
+
+
+
+    private setupFocusTrap(): void {
+        try {
+            this.focusTrap = focusTrap(this.elRef.nativeElement, {
+                clickOutsideDeactivates: true,
+                returnFocusOnDeactivate: true,
+                escapeDeactivates: false
+            });
+        } catch (e) {
+            console.warn('Unsuccessful attempting to focus trap the Combobox.');
         }
     }