+++ library/src/lib/menu/menu-item-addon.directive.ts
@@ -1,0 +1,13 @@
+import { Directive, HostBinding } from '@angular/core';
+
+/**
+ * The directive for menu item addon(for icons).
+ */
+@Directive({
+    selector: '[fd-menu-item-addon], [fdMenuItemAddon]'
+})
+export class MenuItemAddonDirective {
+    /** @hidden */
+    @HostBinding('class.fd-menu__addon-before')
+    fdMenuItemAddonClass: boolean = true;
+}